added json list of groups, thoughts in overview, (hopefully) improved performance by reworking the matching routine
The code's not even close to clean, but I have implemented a JSON list
of groups to check. In short, the script will open a JSON file and
parse it for an array of group.

Added some thoughts to overview.txt re: handling friends in multiple
groups, visualization

Reworked the way we match the member list against the friend list to
streamline it just a bit. Namely, rather than building an array of each
group's member ids, we build an array of the user's friends' ids.

Also added a couple of utility functions (i.e., gtg to check that a var
is set and non-whitespace, and preg_grep_assoc to grep through an array
of associative arrays for a given pattern in a given element of that
assoc).

Added some comments to the code to (hopefully) clarify what's going on.