Creating design and layout
I'm going to add in a folder called "demo." In that folder I will build
out all of how the HTML,CSS, and JavaScript should work. I'm doing it
this way so that I can build styles for the page independent of the
app. That way I don't need you to push files live for me all the time.
Once my design is finished I'll merge the code.